complete working with make

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: matcardo <matcardo@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/04 23:47:30 by matcardo          #+#    #+#             */
-/*   Updated: 2022/06/08 05:03:43 by matcardo         ###   ########.fr       */
+/*   Updated: 2022/06/08 05:22:08 by matcardo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_printf(const char *str, ...);
 int		putstr(char *str);
 int		putnbr(int n);
 int		putnbr_u(unsigned int n);
-int		puthex(unsigned int n, int flag);
+int		puthex(unsigned long int n, int flag);
 int		putptr(unsigned long int n);
 
 #endif
